refactor(dev-center): id on the card itself + shared useScrollToHash hook [preview:none]
Two cleanups on the deep-link-scroll work landed earlier in this PR.

1. DOM id lives on the card itself, no wrapper.

   DeliveryRow + RoadmapCard now take an `id?: string` prop applied to
   their own outer element (along with an unconditional `scroll-mt-24`
   so the row/card always lands BELOW the sticky chrome after a scroll,
   whether or not an id is set — the future caller adding `id` doesn't
   have to remember to ask for the offset).

   - DeliveryTable: dropped the `id` + `scroll-mt-24` I had added to the
     row-separator wrapper. The wrapper stays because it owns the
     `border-b border-ods-border last:border-b-0` styling; the id goes
     directly onto DeliveryRow.
   - RoadmapGrid: dropped the wrapper div entirely (it ONLY existed for
     the id). RoadmapCard is now rendered directly as `<RoadmapCard
     id={`roadmap-${item.id}`} … />` — one less DOM node per card.

2. `useScrollToHash` hook extracted.

   The hash-dispatch effect was identical in delivery-lists.tsx and
   roadmap-view.tsx (poll via rAF until the target element mounts,
   then `scrollElementIntoView`; re-run on `data` ref change + native
   `hashchange`). Now lives in `src/hooks/use-scroll-to-hash.ts` and is
   called as `useScrollToHash(data, { headerOffset: 96 })` from both
   views — same behavior, one definition.
