+++ openframe-frontend-core/src/components/shared/roadmap/roadmap-grid.tsx
@@ -134,20 +134,19 @@ function RoadmapGridSingle({
   return (
     <div className={`grid grid-cols-1 md:grid-cols-2 gap-6 ${showLeftMargin ? 'md:ml-[120px]' : ''}`}>
       {items.map((item) => (
-        // DOM id mirrors the URL anchor `buildDevSectionUrl('roadmap',
-        // <id>)` produces (`#roadmap-<external_id>`). The view's
-        // `hashchange` effect (`roadmap-view.tsx`) reads this and calls
-        // `scrollElementIntoView` on the matching node so chat-card
-        // deep-links land the user on the targeted card.
-        // `scroll-mt-24` keeps the card BELOW the sticky chrome.
-        <div key={item.id} id={`roadmap-${item.id}`} className="scroll-mt-24">
-          <RoadmapCard
-            item={item}
-            userVote={getVote(item.id)}
-            onVote={(voteType) => onVote(item.id, voteType)}
-            isVoting={votingTasks.has(item.id)}
-          />
-        </div>
+        // DOM id + sticky-header scroll offset live ON RoadmapCard's own
+        // outer element (no wrapper div). Anchor mirrors
+        // `buildDevSectionUrl('roadmap', <id>)` → `#roadmap-<external_id>`;
+        // `useScrollToHash` in `roadmap-view.tsx` finds the card by id
+        // and scrolls.
+        <RoadmapCard
+          key={item.id}
+          item={item}
+          id={`roadmap-${item.id}`}
+          userVote={getVote(item.id)}
+          onVote={(voteType) => onVote(item.id, voteType)}
+          isVoting={votingTasks.has(item.id)}
+        />
       ))}
     </div>
   );
