+++ openframe-frontend-core/src/hooks/index.ts
@@ -28,6 +28,11 @@ export * from './use-access-code-integration'
 // OG placeholder URL builder hook (requires host-supplied URL builder)
 export * from './use-og-placeholder'
 
+// Deep-link "scroll to URL hash" after data loads. Pairs with URL
+// composers that emit `?<filter>=<id>#<prefix>-<id>` — the filter
+// narrows the list, the hash scrolls the matching DOM id.
+export * from './use-scroll-to-hash'
+
 // Invisible bot-protection client primitive (honeypot ref + submit-timing).
 // Pairs with the server-safe decision fn in `utils/humanity-signals`.
 export * from './use-humanity-signals'
